Pytest refactor folder structure
