+++ tests/ab_test/aa_tester/unit_test.py
@@ -1,31 +1,9 @@
-import pandas as pd
 import pytest
-
-import sys
-from pathlib import Path
-
-sys.path.append(str(Path(".").absolute().parent))
-
+import pandas as pd
 from hypex import AATest
-from hypex.utils.tutorial_data_creation import create_test_data
-
-
-@pytest.fixture
-def data():
-    return create_test_data(rs=52)
-
-
-@pytest.fixture
-def iterations():
-    return 20
-
-
-@pytest.fixture
-def info_col():
-    return "user_id"
-
 
 def test_aa_simple(data, iterations, info_col):
+    
     model = AATest(target_fields=["pre_spends", "post_spends"], info_cols=info_col)
     res, datas_dict = model.calc_uniform_tests(data, iterations=iterations, pbar=False)
 
@@ -39,7 +17,6 @@ def test_aa_simple(data, iterations, info_col):
     assert all(data.columns) == all(
         datas_dict[0].drop(columns=["group"]).columns
     ), "Columns in the result are not the same as columns in initial data "
-
 
 def test_aa_group(data, iterations, info_col):
     group_cols = "industry"
@@ -61,7 +38,6 @@ def test_aa_group(data, iterations, info_col):
     assert all(data.columns) == all(datas_dict[0].drop(columns=["group"]).columns), (
         "Columns in the result are not " "the same as columns in initial " "data "
     )
-
 
 def test_aa_quantfields(data, iterations, info_col):
     group_cols = "industry"
@@ -86,7 +62,6 @@ def test_aa_quantfields(data, iterations, info_col):
         "Columns in the result are not " "the same as columns in initial " "data "
     )
 
-
 def test_aa_process(data, iterations, info_col):
     group_cols = "industry"
 
@@ -97,11 +72,9 @@ def test_aa_process(data, iterations, info_col):
     )
     model.process(data, iterations=iterations, show_plots=False, pbar=False)
 
-
 def test_group_optimization(data, info_col):
     model = AATest(target_fields=["pre_spends", "post_spends"], info_cols=info_col)
     model.process(data, optimize_groups=True, iterations=5, show_plots=False, pbar=False)
-
 
 def test_unbalanced_groups(data, iterations, info_col):
     test_size = 0.3
